Fixed incorrect return type of currently unused function (just in case it ever gets re-used)
svn-id: r31377

+++ tools/create_lure/create_lure_dat.h
@@ -391,7 +391,7 @@ public:
 		read(&v, sizeof(uint16));
 		return FROM_LE_16(v);
 	}
-	uint16 readLong() {
+	uint32 readLong() {
 		uint32 v;
 		read(&v, sizeof(uint32));
 		return FROM_LE_32(v);
